refactor: split out the llvm context and modules from the builder
Now the builder dose not hold ownership of the llvm module or the context. All
the builder is doing is taking a module and adding the instructions to it.

There is now only one context that is shared between all the modules and the
runner so we are creating and destroying modules multiple times. This was
causing double free errors when trying to delete the llvm execution engine.